% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateAlignments.R
\name{generateAlignments}
\alias{generateAlignments}
\title{Generate Simulated alignements}
\usage{
generateAlignments(algfile, partfile, tree, model, nreps,
  formatin = "sequential", pathFiles = ".",
  pathResults = file.path(pathFiles, "simDataGrouped"),
  raxmlCmd = "raxmlHPC-PTHREADS-SSE3", raxmlArg = "-T7")
}
\arguments{
\item{algfile}{The original alignment file, can be in any format
supported by ape:::read.dna}

\item{partfile}{The partition file (at this stage only RAxML
partition files are supported)}

\item{tree}{A tree file in Newick format}

\item{model}{The model of molecular evolution to be used for
parameter estimation on each parameters (at this stage, only
\dQuote{\code{GTRGAMMA}} and \dQuote{\code{GTRGAMMAI}} are
supported).}

\item{nreps}{Number of replicated simulated datasets to generate.}

\item{formatin}{The alignment format needs to be a format
supported by ape::read.dna.}

\item{pathFiles}{Character string indicating the path where the
data files (\code{algfile}, \code{partfile}, \code{tree}) as well
as \code{seq-gen} are located. Default if current working
directory.}

\item{pathResults}{Character string indicating the path where the
files generated by this function will be written. Default is a
subdirectory of \code{pathFiles} named \code{simDataGrouped}.}

\item{raxmlCmd}{Which RAxML flavor do you want to use? (default
raxmlHPC-PTHREADS-SSE).}

\item{raxmlArg}{Additional arguments to be passed to RAxML.}
}
\value{
Invisibly returns some (possibly) useful location of files
used and generated by this sequence.
}
\description{
This function takes an alignment (with partition information if
needed), a tree, and simulate datasets under a given model
according to the guide tree provided. Each partition is optimized
on the guide tree for all the model parameters (transition rates,
base frequencies, gamma parameter, and proportion of invariants).
}
\details{
This function is used for its side effect, as it generates
replicated simulated datasets.
}
\author{
Francois Michonneau
}

