\name{getParams}
\alias{getParams}
\title{Get alignment characteristics from a RAxML info file}
\usage{
getParams(infoFile)
}
\arguments{
  \item{infoFile}{path (including file name) of the info
  file to be parsed.}
}
\value{
A list that includes various characteristics about the
alignment: \describe{ \item{alpha}{the shape parameter of
the Gamma distribution for the heterogeneity in rates of
evolution.} \item{invar}{the proportion of invariant sites
(empty if the GTRGAMMA model is used)} \item{rateAC,
rateAG, rateAT, rateCG, rateGT}{transition rates for the
GTR model} \item{freqA, freqC, freqG, freqT}{relative
proportions of the these base pairs in the alignement}
\item{nameAlg}{the file name of the alignment file}
\item{treePath}{the file name of the tree with branch
lengths optimized given the parameters of the alignment}
\item{fileNm}{the file name of the info file}
\item{seqlen}{the sequence length} }
}
\description{
This functions parse the output (info file) of an analysis
conducted with RAxML to return the parameters estimated on
an alignment as a list.
}
\details{
Given an \sQuote{info} file generated by RAxML, this
function will return a list that contains all the
parameters characterizing the alignement used. This
function is specifically (and for now, uniquely) designed
to take as an input file, only the RAxML info file
generated after invoking the options \dQuote{-f e -t} while
using the models of molecular evolution GTRGAMMA, and
GTRGAMMAI.

Trying to use this function on other types of info files
generated by RAxML will most likely not work and might give
error messages or not.
}
\author{
Francois Michonneau
}

