\name{splitSimSeq}
\alias{splitSimSeq}
\title{splitSimSeq -- split simulated sequences}
\usage{
splitSimSeq(file, prefix, pathin, pathout)
}
\arguments{
  \item{file}{the path and file name of the simulated file
  to be broken down.}

  \item{prefix}{the prefix for each of the file name}

  \item{pathin}{path (i.e., directory) where the replicates
  for each locus are stored}

  \item{pathout}{path (i.e., directory) where the
  individual replicate for each locus are going to be
  stored.}
}
\value{
TRUE, but really used for its side effect of generating
individual alignment files.
}
\description{
Format Alignment for RAxML input
}
\details{
seq-gen generates all replicates in a single file. This
function reads the file generated by seq-gen and breaks
each replicate into its own file. The files are named by
putting together a prefix (specified by the user), the
string \sQuote{rep} (stands for replicate), the replicate
number, the name of the marker/locus (which should be part
of the file name between an underscore and \sQuote{.phy},
e.g., \code{myalignment_marker.phy.out_simSeq}).
}
\author{
Francois Michonneau
}

