\name{generateBashScript}
\alias{generateBashScript}
\title{Generate a bash script}
\usage{
generateBashScript(path, output, partfile, model, prefix, tree, seed,
  raxmlCmd = "raxmlHPC-PTHREADS-SSE3", raxmlArg = "-T7")
}
\arguments{
  \item{path}{path (i.e., directory) where the concatenated
  datasets are stored.}

  \item{output}{file name (and path if needed) of where the
  script should be written.}

  \item{partfile}{the partition file to be used for the
  replicate (WARNING: this is *NOT* the same as the
  partition file you might have used in your original
  anlysis, instead use the path for the file that was
  generated by \code{\link{finalizeAlignments}}).}

  \item{model}{Model of molecular evolution to be used
  during the likelihood estimation on the replicated
  datasets (should be the same as the one used for the
  original analyses.)}

  \item{prefix}{Prefix used to identify the replicated
  datasets.}

  \item{tree}{If NULL or missing, RAxML will look for the
  \sQuote{best} tree (use the \sQuote{\code{-f d}} RAxML
  algorithm), otherwise a character string indicating the
  file name of a tree to be used as a constraint during the
  likelihood search (use the \sQuote{\code{-f d -g}} RAxML
  algorithm).}

  \item{seed}{Seed to be used to initiate the likelihood
  search in RAxML.}

  \item{raxmlCmd}{Which flavor of RAxML do you want to use?
  (default is HPC-PTHREADS-SSE3 version).}

  \item{raxmlArg}{Additional arguments to be passed to
  RAxML (default "-T7")}
}
\value{
TRUE, but really used for its side effect of generating a
bash script that contains the appropriate command to run a
likelihood search on each of the replicate dataset.
}
\description{
This function generates a bash script that runs RAxML on
each of the replicated dataset.
}
\details{
For now, the bash script generated by this function is not
very elegant (to say the least), but it gets the job done.
}
\author{
Francois Michonneau
}

